Made donating lower priorities work
Basically, the check in lock release only checked if the lock holder has a lock on priority.
It does not check whether the lock of the priority was for the current lock.
Because lock_release method is called by multiple places, the priority for the main thread was revoked if only the priority was locked.

The solution is to make sure that the lock that is being released has waiting threads and as a sanity check ensure that the highest priority waiting thread has greater or equal priority than current thread. The future implementation of multiple priority donation should refactor this code heavily