Attempt to make multiple priority donations work.
The aim is to for a thread to maintain a list of threads that are waiting. These may or may not contains threads that have donated priorities to the thread.
Each thread also maintains a reference to the lock it was trying to hold. This way we can determine the threads that have donated priority to lock holder.